Support touch scroll in register; remove legacy scroll interactions
Just as an experiment, I'd like to compare how standard scrolling
behavior feels in the register view.  I find the current fast and
quantized scroll to be disconcerting, but I recognize that a lot of
thought, and trial and error, have already gone into the existing
custom scroll interactions.

This patch replaces custom scrolling with a ScrolledWindow container
and mostly the default handlers.