Change default column for SX list tree view
As the tree view is based on the GNC_TREE_VIEW, if the pref_name of a
column is "name" it should also be the default sort column. This is
because the default column for GNC_TREE_VIEW is "name" and when saving
state information, 'gnc_tree_view_save_state' will only save non
default values where possible.

In this case if the name column is selected to be default, no entries
are saved in the gcm file and hence the on next load 'next occurrence'
will be used as the default.