Rename the scheduled edit/new/delete menus
There is a conflict when updating the tooltip menu entries for the
scheduled transaction 'edit' menu. To find the appropriate menu entry,
the list of menu entries is searched by name and this entry conflicts
with the main 'edit' menu which has no tooltip and so the scheduled
'edit' menu entry does not get the correct tooltip.

To fix this just change the entries to 'Edit Schedule'/'New Schedule'
and 'Delete Schedule', this also changes them to the same format as the
account menu options.