Bug 798760 - SLR OK button to complete value entry
When in the SLR dialog and you are adding a value to a scheduled
transaction, if you do not complete the entry with a tab or enter but
press the 'OK' button the value changes back to 'Need Value'.

This change also completes the entry with the 'OK' button and will move
to the next value to be entered or if none complete the dialog.