Bug 798760 - SLR change reminder with value entry
If the scheduled transaction state is 'Reminder' and also requires
values which have been entered but the state has not been changed to
'To Create' the values are lost when OK button pressed.

This change adds a check of state to the 'variable_value_changed_cb'
and if 'Reminder' changes it to 'To Create'.