Bug 798025 - SLR Value entry not tied to location
When you enter a value in the SLR tree view, the cell renderer text
entry box gets the focus ready for keyboard entry. In this state, the
tree view can be scrolled underneath by the mouse wheel or using the
vertical scrollbar, the row is still high lighted but still may be
confusing.

This commit disables the scrolling while an entry has focus and also
disables the horizontal scroll bar to be more effective.