Bug 799121 - Parse error on CSV import on MacOS
Use csv_tximp_preview_sep_button_cb solely as a callback, its error
handling doesn't make sense otherwise. All we want at the end of
preview_refresh is to attempt a tokenize.

In addition to not going into an infinite loop with the message box
this presents a more helpful message suggesting that the encoding is
wrong.