Bug 799148 - Reliable crash when saving a modified saved report...
configuration.

The user's saved configurations included some based on a bad
report-id, so handle the hash table returning #f by skipping that configuration.

There's an odd side-effect: The Save Configuration As action doesn't
prepare the dialog to edit the saved configuration name, the user must
find the new configuration and rename it. It works correctly if there
are no discarded configurations.