Bug 799179 - SLR won't allow change from "Reminder" to any other state
This was due to a previous commit to remove empty instances. It was not
realised that the GtkTreeModel needed to be in sync with the GList of
instances as the location in the model was being used to find the
location in the GList.

To fix this, a new column 'PTR' was added to the model and populated
with instances, instance or variable depending on depth and these are
used to the required objects.