Bug 799224 - Import of QIF gets Bug detected during duplicates (partial fix)
If the new-splits object is null, it means the new account tree from
the current import has no splits. Therefore the (apply min|max dates)
will fail. Omitting the date query is a simple fix to prevent crashing.

This is a partial fix because the crash is likely a symptom of another
bug which causes the new account-tree to be empty.