Bug 799258 - Reports calculating net worth incorrectly after stock split
For average-cost and weighted-average: when the exchange-fn is
created, it would formerly consider the whole book foreign currency
conversions to calculate the date-specific conversions. This bug fix
will limit the book conversions to those earlier than the
date-specific conversions.

The book retrieval functions are expensive, therefore the analysis
splitlists are cached in the date-hash hash-table, which is scoped
within the exchange-fn lambda only, and is garbage collected when
exchange-fn is out of scope.

Also: amend relevant commodity-utilities.scm tests

The old weighted-average and average-cost tests were written to
document the (knowingly) incorrect exchanged amounts as originally
calculated pre-2024.

In 2024, the exchange-fn is fixed so that foreign currency
conversions at a particular report date would consider only
transactions up to the relevant report date.

The first AAPL transaction in this test book is on 09/08/2013: a
purchase of 600 AAPL at $36840 total @ $61.40 each.

Therefore the first test, on 20/02/2012 the AAPL->USD exchange rate
conversion should be zero, since there are no AAPL transactions
yet.

Subsequent AAPL test on 20/02/2014 shows the conversion of 1
AAPL to be 307/5 which matches $61.40.

etc.