[dialog-price-edit-db.cpp] Copy the latest price when creating a new entry from a commodity selection
When a price entry is selected, it's contents get copied 1:1 to the newly created price to facilitate faster entry based on existing values (e.g. the type).
When selecting a commodity instead, the user is presented with an empty price entry where only the commodity is set.
IMO the user ends up with the same values (except the price of course) as the latest entry.
So we check whether we have a (latest) entry and use that as the template, so that e.g. the type gets copied, too.
This spares the user some clicks.