Bug 799298 - Shortcut Ctrl-G does not work in the General Journal register for the default date value
The split list in the General Journal register contains the blank
splits of other open registers which the GUI hides in
split-register-load.c. As they are not entered, the entered date value
is 0 and they will be listed before any split with the default date
value. If such a split is used, nothing will happen.

To fix this use the same test to ignore these blank splits.