Bug 798822 - Move to blank transaction
If the move to blank action is used when a register is displayed in
REG_STYLE_JOURNAL, the resulting highlighted cell is the split 'action'
cell. In all other register styles it is the 'date' cell so test for
the register style and decrement the vcell_loc.virt_row by one to move
to the 'date' cell.