Bug 799320 - GNUCash Immediately Exits on Startup
Some excellent and insightful debugging by the reporter found that the cause was specifying
a locale as en_US.utf8 instead of en_US.utf-8. GnuLib's locale functions and so Guile's are
apparently quite picky on Win32.

To address this move reading the environment file to before we set the mac and Windows
locales and ensure that UTF/UCS and ISO-8859 character sets are specified with minus
separators on Windows. I tested on macOS and didn't experience the crash.