Bug 799047 - AutoComplete Only Considers Visible Transactions
In the past, all transactions were used on the first load of the
register and this is when the autocomplete would be populated, on
subsequent loads it would be a filtered list. With the change to delay
loading, the first load is empty and the autocomplete is now being
populated by the filtered list.

To fix this, make a copy of the original query and compare this query
to the query used for refresh. When different, pass a second list of
splits to the register to populate the autocomplete lists.