Correctly handle uncommitted edits when opening a file from history.
By calling gnc_main_window_finish_pending as is done in all other
execution paths that close the session. That allows an open edit to be
committed or cancelled before the user is prompted to save the file
instead of after when it's too late. It also allows the user to cancel closing the session at that point if there's more work to be done on the transaction.