Bug 799399 - Windows Keypad decimal locale error
On Windows, the keypad decimal isn't a comma in dialogs that use the
GncAmountEdit widget for locales that calls for it.

This appears to be a problem with Gtk so instead do the entry update
in the key press call back instead of trying to change the event and
passing it on.