Warn and don't try to balance a transaction with a too-small imbalance.
In a book using trading accounts, if the user sets an account's
minimum fraction to be smaller than the commodity's and makes use of
the extra fraction, GnuCash can't create a correct trading split
because that uses the commodity fraction. Instead of offering to
balance the transaction, put up an error dialog explaing that the
transaction can't be balanced.

Ref Bug 799432 (https://bugs.gnucash.org/show_bug.cgi?id=799423)