Change the gnc_dialog_run function
Rename the gnc_dialog_run function to gnc_warning_dialog_run as it uses
the preference warning settings and make it modal and pause the main
loop so it can return a value.

Create a new gnc_dialog_run along the lines of above and change only
references to gtk_dialog_run used to create message dialogs to this
function.

This may need a rethink but for now is OK if only temporary, eventually
the widget used will be a GtkAlertDialog.