Change the doc link GtkTreeview to a GtkColumnview
This is just a test to see what is involved to change a GtkTreeview to
a GtkColumnview.

I am not sure if all is correct so for clarity, the column view is in
separate files and in long hand but some functions should be able to be
refactored. The other option would be to do a lot for of the column
view in the ui file.