Deescalate needed SQL privileges for MySQL / Mariadb
This commit drops the need to be able to use the system database "mysql"
which is not accessible to "ordinary" db users, normally.

At least in recent MySQL / MariaDB version it is not necessary to select
a database to issue a "CREATE DATABASE" command, and it is also not
necessary to first switch away from the selected data-base in order to
drop it.

Another approach could be to use the `information_schema` DB instead of
the `mysql` as this would not require super-user privileges.

With this patch it is possible to use MySQL/MariaDB as backend with more
or less ordinary privileges for the the connecting user -- except that
it needs to have CREATE privileges. This is bad from a security point of
view. With MySQL / MariaDB this can be "fixed" to some extend by
granting CREATE on a table-name wildcard, for example:

GRANT CREATE ON `gnucash\_%`.* TO 'gnucash'@'%' IDENTIFIED BY '********';

IMHO, a better approach would be to at least optionally stop using DROP
/ CREATE DATABASE and instead simple drop all the tables of a given
database which would achieve the same goal but with less privilege
escalation.

Signed-off-by: Claus-Justus Heine <himself@claus-justus-heine.de>