Bug 799363 - When selecting "load" two layer ...
The Back and Forward buttons now work correctly.
Added pressedBACK and pressedFWD to _gnc_html_history
because gnc_html_history_append needs to know if a button
has been pressed.

I had to remove the Back and Forward options from the right-click
menu because they interfere with the buttons' functionality.

Clicking and selecting Load clears the existing Forward order.
I can't change it because it's JavaScript inside the HTML.